pyproject.toml
[tool.poetry]
name = "lhw27"
version = "0.1.0"
description = ""
authors = ["DmitriiYaganin <dmitrij.yaganin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
psycopg2 = "^2.9.5"
pillow = "^9.3.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
